+++ server/api/src/test/java/com/twb/pokerapp/testutils/game/turn/impl/OptimisticTurnHandler.java
@@ -6,13 +6,14 @@ import com.twb.pokerapp.testutils.game.turn.TurnHandler;
 import com.twb.pokerapp.web.websocket.message.server.payload.PlayerTurnDTO;
 import org.springframework.messaging.simp.stomp.StompHeaders;
 
+import java.math.BigDecimal;
 import java.util.Arrays;
 
 import static com.twb.pokerapp.testutils.game.turn.TurnHandler.sendPlayerAction;
 
 // Will Bet or Call when it can. It will not check or fold
 public class OptimisticTurnHandler implements TurnHandler {
-    private static final double DEFAULT_BET_AMOUNT = 10d;
+    private static final BigDecimal DEFAULT_BET_AMOUNT = BigDecimal.valueOf(10);
 
     @Override
     public void handle(AbstractTestUser user, StompHeaders headers, PlayerTurnDTO playerTurn) {
